chore - rm rotate handle events
